Switched to using `streamlit` as the main data visualization library
`streamlit` contains *a lot* of functionality and the resulting
application is a web application. It has a set of widgets that can
accept input which can then be used to query data.

Additionally, `streamlit` encompasses multiple data visualization
libraries such as `bokeh`, `altair`, and `matplotlib`.

Given that there may not be a need for explicitly sending the data to
the front-end, we may be able to completely forgo having to generate and
parse a `.json` file.